Fixing some color conversions
- Manually changed from arantius

+++ Catalano.Image/src/Catalano/Imaging/Concurrent/Filters/HSLFiltering.java
@@ -258,7 +258,7 @@ public class HSLFiltering implements IApplyInPlace{
                     int g = share.fastBitmap.getGreen(i, j);
                     int b = share.fastBitmap.getBlue(i, j);
                     
-                    float[] hsl = ColorConverter.RGBtoHSL(r, g, b);
+                    double[] hsl = ColorConverter.RGBtoHSL(r, g, b);
                     
                     // check HSL values
                     if (
